A lot of error handling
Errors 403 and 410 should no longer crash the bot. And the user should receive a message if a song was unable to play due to one of these errors.

Error ERRCON, where the playback stops in the middle of the song is unhandled, due to inability to reproduce.

A fork of the ytdl library "ytdl-code-as" is being used to get the audio buffer stream, as it seems to prevent some errors (such ass ERRCONN).

Inputs outside of the scope have been handled for skip and queue commands.