Add loop and loopqueue commands
Fix issue where on error 403 with loop enabled, the player added more copies of the song into the queue. This was caused because during an error, the player is in a state of idle, which caused the error callback function to access that method.

Update queue to show whether loop is turned on.

Currently loopqueue simply appends the last song at the end of the queue. Not very elegant, but works.