Clear queue on disconnect & disconnect on IDLE
If the bot is left alone in a voice chat, disconnect after 4 minutes of being alone.

If the bot is forcefully disconnected by a user, clear the queue after 5 seconds. If instead it was forcefully disconnected from that VC, but moved into a different VC, use those 5 seconds to listen for the disconnect and keep the queue if a reconnection was made.